requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
anyio==4.12.1
attrs==26.1.0
beautifulsoup4==4.14.3
bidict==0.23.1
certifi==2026.1.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
lxml==6.1.0
multidict==6.7.1
packaging==26.1
pluggy==1.6.0
propcache==0.4.1
pyflakes==3.4.0
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-engineio==3.14.2
python-socketio==4.6.1
python-telegram-bot==22.6
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
wsproto==1.3.2
yarl==1.23.0
